% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data-to-txt.R
\name{breakBeaconData}
\alias{breakBeaconData}
\title{Extracting connections per-child}
\usage{
breakBeaconData(beaconDat)
}
\arguments{
\item{beaconDat}{Takes input from binningBeaconData function}
}
\value{
a list of connections per child
}
\description{
beacon Data is given in overall overview, here, the information is broken into individual chunks, representing individual children and their
}
