% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data-to-txt.R
\name{get.txt.files}
\alias{get.txt.files}
\title{Apply csv to txt script on one group}
\usage{
get.txt.files(binsize, minlength, minlag, group)
}
\arguments{
\item{binsize}{size of the window or bin to be binned over}

\item{minlength}{indicates the minimum length of an interaction}

\item{minlag}{indicates how much lag between interactions is needed/wanted}

\item{group}{depending on how txt are supposed to be named, this is an option
to group data by location, provenance or other info}
}
\value{
.txt files in the folder ~/fluiditypilot/data/textdata_",group,"/ ; also returns a list with two elements
\enumerate{
\item list of interaction lengths; 2) vector with info on how long the recordings are
}
}
\description{
Function to collect info on interaction lengths.
}
