% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data-to-txt.R
\name{getOutputStartEnds}
\alias{getOutputStartEnds}
\title{Turn raw data into format for txt as needed in jupyter notebook}
\usage{
getOutputStartEnds(brokenfile, minlength, minlag)
}
\arguments{
\item{brokenfile}{as from breakBeaconData function}

\item{minlength}{indicates the minimum length of an interaction}

\item{minlag}{indicates how much lag between interactions is needed/wanted}
}
\value{
matrix with tagX, tagY, start, end
}
\description{
Turn raw data into format for txt as needed in jupyter notebook
}
