% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data-to-txt.R
\name{newstart}
\alias{newstart}
\title{Find all newstart moments
If tapplied, it accounts for newstart moments depending on tagY (or whichever tapply dependence)}
\usage{
newstart(x, minlength, minlag)
}
\arguments{
\item{x}{Depends on how the function is tapplied, which ever subset of data is fed. Data needs to be formated as the output from the binningBeaconData function}

\item{minlength}{indicates the minimum length of an interaction}

\item{minlag}{indicates how much lag between interactions is needed/wanted}
}
\value{
filtered data, where minlength and minlag are applied on all connections
}
\description{
Find all newstart moments
If tapplied, it accounts for newstart moments depending on tagY (or whichever tapply dependence)
}
