% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data-to-txt.R
\name{turnListintoOutput}
\alias{turnListintoOutput}
\title{Extract and format start and end info}
\usage{
turnListintoOutput(listelement)
}
\arguments{
\item{listelement}{as from the output of newstart helper-function}
}
\value{
a cbind startinfo, endinfo combination
}
\description{
Extract and format start and end info
}
