% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data-to-txt.R
\name{write.output}
\alias{write.output}
\title{Write txt for python script}
\usage{
write.output(output, binsize, minlag, minlength, filename, group)
}
\arguments{
\item{output}{the object that you want to store as txt}

\item{binsize}{size of the window or bin to be binned over}

\item{minlag}{indicates how much lag between interactions is needed/wanted}

\item{minlength}{indicates the minimum length of an interaction}

\item{filename}{filename that indicates which data we are working with (e.g., School5_morning or similar)}

\item{group}{depending on how txt are supposed to be named, this is an option to group data by location, provenance or other info}
}
\value{
.txt file in the folder ~/fluiditypilot/data/textdata_",group,"/
}
\description{
Exports txt file needed for the python scripts/ jupyter notebook to run
Keeps track of binsize, minlag, minlength parameter when naming the output files
}
