% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data-to-txt.R
\name{binningBeaconData}
\alias{binningBeaconData}
\title{Binning beacon or beagle data}
\usage{
binningBeaconData(dataCSV, binsize)
}
\arguments{
\item{dataCSV}{csv in format: tagX, tagY, time}

\item{binsize}{can be anything >1}
}
\value{
returns the data as X, Y, bin indicator (e.g., bin indicator = 1 for all connections made during bin 1)
}
\description{
Binning beacon/ beagle data (or any csv in format: tagX, tagY, time) and applying by default a default 50\% minimum signal per bin on the data to count the bin as signal.
}
